# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /closeouts:
    get:
      operationId: listCloseouts
      tags: [closeouts]
      summary: List closeouts for the current tenant
      responses:
        "200":
          description: Closeouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Closeout"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCloseout
      tags: [closeouts]
      summary: Submit a VoiceConnect closeout draft (routes to Pending Review)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /closeouts/{id}:
    patch:
      operationId: updateCloseout
      tags: [closeouts]
      summary: Edit a closeout draft (before approval)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseoutUpdate"
      responses:
        "200":
          description: Updated closeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /closeouts/{id}/approve:
    post:
      operationId: approveCloseout
      tags: [closeouts]
      summary: Approve a closeout (idempotent) — posts labor/materials, deducts inventory once, hands off to billing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved closeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /closeouts/{id}/send-back:
    post:
      operationId: sendBackCloseout
      tags: [closeouts]
      summary: Send a closeout back to the technician with a reason
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseoutReviewInput"
      responses:
        "200":
          description: Sent back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Closeout:
      type: object
      required:
        - id
        - tenantId
        - workOrderId
        - technicianId
        - submittedAt
        - transcript
        - transcriptLanguage
        - aiSummary
        - workPerformed
        - materialsDetected
        - laborSuggested
        - missingInfo
        - customerUpdateText
        - billingLines
        - portalUpdateText
        - status
      properties:
        id: { type: string }
        tenantId: { type: string }
        workOrderId: { type: string }
        technicianId: { type: string }
        submittedAt: { type: string, format: date-time }
        transcript: { type: string }
        transcriptLanguage: { type: string }
        translatedSummary: { type: ["string", "null"] }
        aiSummary: { type: string }
        workPerformed: { type: string }
        materialsDetected:
          type: array
          items: { type: string }
        laborSuggested: { type: string }
        returnTripReason: { type: ["string", "null"] }
        quoteNotes: { type: ["string", "null"] }
        missingInfo:
          type: array
          items: { type: string }
        customerUpdateText: { type: string }
        billingLines:
          type: array
          items: { type: string }
        portalUpdateText: { type: string }
        status: { type: string }
        reviewedBy: { type: ["string", "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        reviewNote: { type: ["string", "null"] }
    CloseoutInput:
      type: object
      required: [workOrderId, technicianId]
      properties:
        workOrderId: { type: string, minLength: 1 }
        technicianId: { type: string, minLength: 1 }
        transcript: { type: string }
        transcriptLanguage: { type: string }
        translatedSummary: { type: string }
        aiSummary: { type: string }
        workPerformed: { type: string }
        materialsDetected:
          type: array
          items: { type: string }
        laborSuggested: { type: string }
        returnTripReason: { type: string }
        quoteNotes: { type: string }
        missingInfo:
          type: array
          items: { type: string }
        customerUpdateText: { type: string }
        billingLines:
          type: array
          items: { type: string }
        portalUpdateText: { type: string }
    CloseoutUpdate:
      type: object
      properties:
        transcript: { type: string }
        transcriptLanguage: { type: string }
        translatedSummary: { type: string }
        aiSummary: { type: string }
        workPerformed: { type: string }
        materialsDetected:
          type: array
          items: { type: string }
        laborSuggested: { type: string }
        returnTripReason: { type: string }
        quoteNotes: { type: string }
        missingInfo:
          type: array
          items: { type: string }
        customerUpdateText: { type: string }
        billingLines:
          type: array
          items: { type: string }
        portalUpdateText: { type: string }
    CloseoutReviewInput:
      type: object
      properties:
        reason: { type: string }
...